+++ database/protocol/dialect/mysql/src/main/java/org/apache/shardingsphere/database/protocol/mysql/packet/command/query/text/MySQLTextResultSetRowPacket.java
@@ -17,15 +17,21 @@
 
 package org.apache.shardingsphere.database.protocol.mysql.packet.command.query.text;
 
+import com.google.common.io.ByteStreams;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.database.protocol.mysql.packet.MySQLPacket;
 import org.apache.shardingsphere.database.protocol.mysql.payload.MySQLPacketPayload;
+import org.apache.shardingsphere.infra.exception.generic.UnknownSQLException;
 import org.apache.shardingsphere.infra.util.datetime.DateTimeFormatterFactory;
 
+import java.io.IOException;
 import java.math.BigDecimal;
+import java.sql.Clob;
+import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Collection;
 
@@ -71,6 +77,15 @@ public final class MySQLTextResultSetRowPacket extends MySQLPacket {
             payload.writeBytesLenenc((boolean) data ? new byte[]{1} : new byte[]{0});
         } else if (data instanceof LocalDateTime) {
             payload.writeStringLenenc(formatLocalDateTime((LocalDateTime) data));
+        } else if (data instanceof LocalTime) {
+            payload.writeStringLenenc(formatLocalTime((LocalTime) data));
+        } else if (data instanceof Clob) {
+            try {
+                // TODO Verify the correct approach for this in MySQL.
+                payload.writeBytesLenenc(ByteStreams.toByteArray(((Clob) data).getAsciiStream()));
+            } catch (final IOException | SQLException ex) {
+                throw new UnknownSQLException(ex);
+            }
         } else {
             payload.writeStringLenenc(data.toString());
         }
@@ -82,6 +97,24 @@ public final class MySQLTextResultSetRowPacket extends MySQLPacket {
             return DateTimeFormatterFactory.getDatetimeFormatter().format(value);
         }
         StringBuilder result = new StringBuilder(DateTimeFormatterFactory.getDatetimeFormatter().format(value)).append('.');
+        String microsecondsText = String.format("%06d", nanos / 1000);
+        int endIndex = microsecondsText.length();
+        while (endIndex > 0 && '0' == microsecondsText.charAt(endIndex - 1)) {
+            endIndex--;
+        }
+        if (0 == endIndex) {
+            return result.substring(0, result.length() - 1);
+        }
+        result.append(microsecondsText, 0, endIndex);
+        return result.toString();
+    }
+    
+    private String formatLocalTime(final LocalTime value) {
+        int nanos = value.getNano();
+        if (0 == nanos) {
+            return DateTimeFormatterFactory.getTimeFormatter().format(value);
+        }
+        StringBuilder result = new StringBuilder(DateTimeFormatterFactory.getTimeFormatter().format(value)).append('.');
         String microsecondsText = String.format("%06d", nanos / 1000);
         int endIndex = microsecondsText.length();
         while (endIndex > 0 && '0' == microsecondsText.charAt(endIndex - 1)) {