+++ kernel/sql-federation/compiler/src/main/java/org/apache/shardingsphere/sqlfederation/compiler/sql/ast/converter/segment/expression/ExpressionConverter.java
@@ -92,7 +92,7 @@ public final class ExpressionConverter {
             return Optional.empty();
         }
         if (segment instanceof LiteralExpressionSegment) {
-            return LiteralExpressionConverter.convert((LiteralExpressionSegment) segment);
+            return LiteralExpressionConverter.convert((LiteralExpressionSegment) segment, null);
         }
         if (segment instanceof CommonExpressionSegment) {
             // TODO